requirements.txt
opencv-python==4.10.0.84
Pillow==10.4.0
imagehash==4.3.1
requests==2.32.3
tqdm==4.66.5
python-dotenv==1.0.1
numpy==1.26.4
